+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-fixture/src/test/java/org/apache/shardingsphere/test/integration/junit/runner/parallel/impl/CaseParallelRunnerExecutor.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.test.integration.engine.junit.parallel.impl;
+package org.apache.shardingsphere.test.integration.junit.runner.parallel.impl;
 
 import org.apache.shardingsphere.infra.executor.kernel.thread.ExecutorServiceManager;
-import org.apache.shardingsphere.test.integration.engine.junit.parallel.ParallelRunnerExecutor;
-import org.apache.shardingsphere.test.integration.engine.param.model.ParameterizedArray;
+import org.apache.shardingsphere.test.integration.junit.runner.parallel.ParallelRunnerExecutor;
+import org.apache.shardingsphere.test.integration.junit.param.model.ParameterizedArray;
 
 import java.util.Collection;
 import java.util.LinkedList;
